+++ allensdk/core/sitk_utilities.py
